40-m1-implement-a-folder-structure: Folder structure has been fixed. Test has been added to package.json. CMCD end to end js has been added.
