+++ src/inputValidator/index.js
@@ -1,0 +1,2 @@
+export { default as jsonIsValid } from './jsonValidator.js';
+export { default as queryValidator } from './queryValidator.js';
