+++ src/keyValueValidator/index.js
@@ -1,0 +1,1 @@
+export { default as keyValValidator } from './keysValueValidator.js';
