+++ src/parser/parseHeaderToJson.js
@@ -1,0 +1,5 @@
+const parseHeaderToJson = (queryString) => {
+  console.log(queryString);
+};
+
+export default parseHeaderToJson;
