+++ src/tests/index.js
@@ -1,0 +1,7 @@
+import { testQueryValidator, testJsonValidator } from './inputValidator/index.js';
+
+console.log('--------- Test queryValidator---------');
+testQueryValidator();
+console.log('\n');
+console.log('--------- Test jsonValidator---------');
+testJsonValidator();
