+++ src/tests/inputValidator/index.js
@@ -1,8 +1,2 @@
-import { testJsonValidator } from './testJsonValidator.js';
-import { testQueryValidator } from './testQueryValidator.js';
-
-console.log('--------- Test queryValidator---------');
-testQueryValidator();
-console.log('\n');
-console.log('--------- Test jsonValidator---------');
-testJsonValidator();
+export { default as testJsonValidator } from './testJsonValidator.js';
+export { default as testQueryValidator } from './testQueryValidator.js';
